Support reading dependency-groups from pyproject.tomls with no project
(or legacy tool.uv.workspace).

This cleaves out a dedicated SourcedDependencyGroups type based on RequiresDist
but with only the DependencyGroup handling implemented. This allows `uv pip`
to read `dependency-groups` from pyproject.tomls that only have that table
defined, per PEP 735, and as implemented by `pip`.

However we want our implementation to respect various uv features when they're
available:

* `tool.uv.sources`
* `tool.uv.index`
* `tool.uv.dependency-groups.mygroup.requires-python` (#13735)

As such we want to opportunistically detect "as much as possible" while doing
as little as possible when things are missing. The issue with the old RequiresDist
path was that it fundamentally wanted to build the package, and if `[project]`
was missing it would try to desperately run setuptools on the pyproject.toml
to try to find metadata and make a hash of things.

At the same time, the old code also put in a lot of effort to try to pretend that
`uv pip` dependency-groups worked like `uv` dependency-groups with defaults
and non-only semantics, only to separate them back out again. By explicitly
separating them out, we confidently get the expected behaviour.

Note that dependency-group support is still included in RequiresDist, as some
`uv` paths still use it. It's unclear to me if those paths want this same
treatment -- for now I conclude no.

Fixes #13138